Reconfigure project to allow packaging
This includes setting up electron-compile correctly, and getting rid
of electron-prebuild-compile.

A big problem right now is that for some reason, the packaged app is
unreasonably huge. 400 mb for the prototype??? Current speculation is
that the app currently needs to transpile ES6 into ES5, therefore the
whole electron module is required to be packaged, which contributes
at least 250mb. This also explains why packaing with --no-prune flag
will not work, because the app couldn't transpile without electron
package source data.

Need to figure out a way to transpile the code prior to packaging..
Maybe look into electron-forge, and hope that it takes care of
everything?